Switched time in queue histogram to an exponential one
