+++ app/carbonapi/app.go
@@ -56,136 +56,6 @@ type App struct {
 	prometheusMetrics PrometheusMetrics
 }
 
-// PrometheusMetrics are metrix exported via /metrics endpoint for Prom scraping
-type PrometheusMetrics struct {
-	Requests     prometheus.Counter
-	Responses    *prometheus.CounterVec
-	DurationsExp prometheus.Histogram
-	DurationsLin prometheus.Histogram
-	TimeInQueue  prometheus.Histogram
-}
-
-func newPrometheusMetrics(config cfg.API) PrometheusMetrics {
-	return PrometheusMetrics{
-		Requests: prometheus.NewCounter(
-			prometheus.CounterOpts{
-				Name: "http_requests_total",
-				Help: "Count of HTTP requests",
-			},
-		),
-		Responses: prometheus.NewCounterVec(
-			prometheus.CounterOpts{
-				Name: "http_responses_total",
-				Help: "Count of HTTP responses, partitioned by return code and handler",
-			},
-			[]string{"code", "handler"},
-		),
-		DurationsExp: prometheus.NewHistogram(
-			prometheus.HistogramOpts{
-				Name:    "http_request_duration_seconds_exp",
-				Help:    "The duration of HTTP requests (exponential)",
-				Buckets: prometheus.ExponentialBuckets((50 * time.Millisecond).Seconds(), 2.0, 20),
-			},
-		),
-		DurationsLin: prometheus.NewHistogram(
-			prometheus.HistogramOpts{
-				Name:    "http_request_duration_seconds_lin",
-				Help:    "The duration of HTTP requests (linear)",
-				Buckets: prometheus.LinearBuckets(0.0, (50 * time.Millisecond).Seconds(), 40), // Up to 2 seconds
-			},
-		),
-		TimeInQueue: prometheus.NewHistogram(
-			prometheus.HistogramOpts{
-				Name: "time_in_queue",
-				Help: "Time a request to backend spends in queue, in ms",
-				Buckets: prometheus.LinearBuckets(
-					config.Zipper.Common.Monitoring.TimeInQueueHistogram.Start,
-					config.Zipper.Common.Monitoring.TimeInQueueHistogram.BucketSize,
-					config.Zipper.Common.Monitoring.TimeInQueueHistogram.BucketsNum),
-			},
-		),
-	}
-}
-
-// TODO (grzkv): Remove from global scope
-var apiMetrics = struct {
-	// Total counts across all request types
-	Requests  *expvar.Int
-	Responses *expvar.Int
-	Errors    *expvar.Int
-
-	Goroutines    expvar.Func
-	Uptime        expvar.Func
-	LimiterUse    expvar.Func
-	LimiterUseMax expvar.Func
-
-	// Despite the names, these only count /render requests
-	RenderRequests        *expvar.Int
-	RequestCacheHits      *expvar.Int
-	RequestCacheMisses    *expvar.Int
-	RenderCacheOverheadNS *expvar.Int
-
-	FindRequests        *expvar.Int
-	FindCacheHits       *expvar.Int
-	FindCacheMisses     *expvar.Int
-	FindCacheOverheadNS *expvar.Int
-
-	MemcacheTimeouts expvar.Func
-
-	CacheSize  expvar.Func
-	CacheItems expvar.Func
-}{
-	Requests:  expvar.NewInt("requests"),
-	Responses: expvar.NewInt("responses"),
-	Errors:    expvar.NewInt("errors"),
-
-	// TODO: request_cache -> render_cache
-	RenderRequests:        expvar.NewInt("render_requests"),
-	RequestCacheHits:      expvar.NewInt("request_cache_hits"),
-	RequestCacheMisses:    expvar.NewInt("request_cache_misses"),
-	RenderCacheOverheadNS: expvar.NewInt("render_cache_overhead_ns"),
-
-	FindRequests: expvar.NewInt("find_requests"),
-
-	FindCacheHits:       expvar.NewInt("find_cache_hits"),
-	FindCacheMisses:     expvar.NewInt("find_cache_misses"),
-	FindCacheOverheadNS: expvar.NewInt("find_cache_overhead_ns"),
-}
-
-// TODO (grzkv): Remove from global scope
-var zipperMetrics = struct {
-	FindRequests *expvar.Int
-	FindErrors   *expvar.Int
-
-	RenderRequests *expvar.Int
-	RenderErrors   *expvar.Int
-
-	InfoRequests *expvar.Int
-	InfoErrors   *expvar.Int
-
-	Timeouts *expvar.Int
-
-	CacheSize  expvar.Func
-	CacheItems expvar.Func
-
-	CacheMisses *expvar.Int
-	CacheHits   *expvar.Int
-}{
-	FindRequests: expvar.NewInt("zipper_find_requests"),
-	FindErrors:   expvar.NewInt("zipper_find_errors"),
-
-	RenderRequests: expvar.NewInt("zipper_render_requests"),
-	RenderErrors:   expvar.NewInt("zipper_render_errors"),
-
-	InfoRequests: expvar.NewInt("zipper_info_requests"),
-	InfoErrors:   expvar.NewInt("zipper_info_errors"),
-
-	Timeouts: expvar.NewInt("zipper_timeouts"),
-
-	CacheHits:   expvar.NewInt("zipper_cache_hits"),
-	CacheMisses: expvar.NewInt("zipper_cache_misses"),
-}
-
 // New creates a new app
 func New(config cfg.API, logger *zap.Logger, buildVersion string) (*App, error) {
 	if len(config.Backends) == 0 {
