+++ app/carbonapi/http_handlers.go
@@ -127,7 +127,7 @@ func initHandlers(app *App) http.Handler {
 	return r
 }
 
-func writeResponse(w http.ResponseWriter, b []byte, format string, jsonp string, ctx context.Context) {
+func writeResponse(ctx context.Context, w http.ResponseWriter, b []byte, format string, jsonp string) {
 
 	w.Header().Set("X-Carbonapi-UUID", util.GetUUID(ctx))
 	switch format {
@@ -283,7 +283,7 @@ func (app *App) renderHandler(w http.ResponseWriter, r *http.Request) {
 
 		if err == nil {
 			apiMetrics.RequestCacheHits.Add(1)
-			writeResponse(w, response, format, jsonp, ctx)
+			writeResponse(ctx, w, response, format, jsonp)
 			accessLogDetails.FromCache = true
 			return
 		}
@@ -506,7 +506,7 @@ func (app *App) renderHandler(w http.ResponseWriter, r *http.Request) {
 		body = png.MarshalSVGRequest(r, results, template)
 	}
 
-	writeResponse(w, body, format, jsonp, ctx)
+	writeResponse(ctx, w, body, format, jsonp)
 
 	if len(results) != 0 {
 		tc := time.Now()
