+++ app/carbonzipper/metrics.go
@@ -99,11 +99,11 @@ func NewPrometheusMetrics(config cfg.Zipper) *PrometheusMetrics {
 		TimeInQueue: prometheus.NewHistogram(
 			prometheus.HistogramOpts{
 				Name: "time_in_queue",
-				Help: "Time a request spends in queue, ms",
-				Buckets: prometheus.LinearBuckets(
-					config.Monitoring.TimeInQueueHistogram.Start,
-					config.Monitoring.TimeInQueueHistogram.BucketSize,
-					config.Monitoring.TimeInQueueHistogram.BucketsNum),
+				Help: "Time a request spends in queue (exponential), ms",
+				Buckets: prometheus.ExponentialBuckets(
+					config.Monitoring.TimeInQueueExpHistogram.Start,
+					config.Monitoring.TimeInQueueExpHistogram.BucketSize,
+					config.Monitoring.TimeInQueueExpHistogram.BucketsNum),
 			},
 		),
 	}
